#50 refactor: SchedulePreview のロジックを index.hooks.ts へ分離
component 本体に store 3種の購読・buildSchedule 呼出が直書きされており、React 実装規約のロジック分離ルールに沿っていなかった。useSchedulePreview として切り出し、component は JSX のみになった。
